requirements.txt
openslide-python
opencv-python
numpy
torch
torchvision
transformers
tqdm
Pillow